Add schemas for IoJ Appeals
This commit adds schemas for creating, updating and getting
an IoJ Appeal. It mostly captures the existing data that
users are required to enter as part of creating an appeal
through MAAT, but changes this somewhat to more clearly
define who is deciding on the appeal (caseworker of judge)
rather than duplicating data through the existing appeal
set-up result.