Remove IoJ Appeal Status field
This commit removes the `appealStatus` field from the
`IojAppealMetadata` schema, following a recent business
decision that for IoJ Appeals specifically, the status
field is not needed. When submitting an IoJ Appeal,
caseworkers already have all of the information that
is necessary to submit the appeal, so there is no value
in saving them with an `IN_PROGRESS` state.