Remove appealId and legacyAppealId from appeal metadata
This commit removes `appealId` and `legacyAppealId` from
the `iojAppealMetadata` schema. The new schema only needs
to accommodate for appeal creation (no updates) and both
ids will not yet be known during the process of _creating_
the appeal; only when the appeal has been created will there
be an `appealId` and possibly `legacyAppealId`, both of
which are defined separately in the response schema.