Add applicationId and legacyApplicationId to iojAppealMetadata schema
This commit adds the `applicationId` and `legacyApplicationId`
to the `iojAppealMetadata` schema. It also removes the requirement
for either one of them to exist in the request (as only one of them
will exist but we cannot guarantee which one, so this should be
validated as part of the service consuming the request).