Update ApiGetIojAppealResponse schema
This commit updates the _ApiGetIojAppealResponseSchema_
to add `caseManagementUnitId` (which is being stored in
the new database and is required by MAAT but not currently
being returned) and removes `appealId` as a required field
(in the cases where a legacy IoJ Appeal is returned, the
`appealId` will _not_ be present and would fail this
requirement). One of `appealId` and `legacyAppealId` will
be set at the minimum, but which one cannot be known for
the purposes of this schema design.