Remove applicationId from IojAppealMetadata schema
This commit removes `applicationId` from the _IojAppealMetadata_
schema. This field will be useful when applications are created
through a new service (and not via MAAT), but this isn't yet
on the horizon so for now we will continue to use
`legacyApplicationId` and review where we stand at such time as
applications are considered to be created via a new service.