Switch CircleCI to use full checkout
This commit switches the build and test step of our
CircleCI CI pipeline to use the `full` git checkout
mode rather than `blobless` git checkout, which became
the default for pipelines on November 3rd 2025 and has
been rolled out to orgs on a gradual basis. This has
affected Crime Commons today as SonarQube is now failing
to run.

As per the [release notes from CircleCI](https://circleci.com/changelog/default-method-used-to-checkout-code-from-your-repository-is-changing-on-nov/),
switching back to the old git checkout method should fix
this issue for SonarQube.